feat: `signature_types` application arg.
This PR adds a new argument to the app and layers, `signature_types`.

Types in this collection are added to `signature_namespace` using the type's
`__name__` attribute.

The objective is to provide a nicer interface when adding names to the
signature namespace without modifying the type name, e.g.:
`signature_namespace={"Model": Model}` is equivalent to
`signature_types=[Model]`.

The implementation makes it an error to supply a type in `signature_types`
that has a value for `__name__` already in the signature namespace.

It will also throw an error if an item in `signature_types` has no `__name__`
attribute.